fix(build): split Unix-only syscall sites into build-tag-gated files for Windows native build (TASK-025)
- daemon_unix.go (!windows): full setupSignalHandlers with SIGUSR2 + SIGHUP
- daemon_windows.go (windows): stub setupSignalHandlers with SIGTERM/Interrupt only
- cli_unix.go (!windows): setSetsid sets Setsid=true; sendForceTriggerSignal sends SIGUSR2
- cli_windows.go (windows): setSetsid uses CREATE_NEW_PROCESS_GROUP; sendForceTriggerSignal uses HTTP timer trigger
- daemon.go: remove setupSignalHandlers body and os/signal import
- cli.go: replace inline SysProcAttr/SIGUSR2 usages with helper calls; remove syscall import

go build/vet/test all pass on Windows. Linux/macOS behavior unchanged via build tags.
